post view implemented but wrongly structured
