+++ Header.module.css
@@ -43,30 +43,24 @@
 } */
 
 
-h1 {
+h1{
   align-items: left;
   display: flex;
 }
 
 a {
-  color: #273f17;
-  ;
-  align-items: right;
-  gap: 2px;
-  margin-right: 20px;
+color:  #273f17;;
+align-items: right;
+gap: 2px;
+margin-right: 20px;
 }
 
 a:hover {
   background-color: #d8e2dc;
   color: #273f17;
 }
-
 a.active {
   background-color: magenta;
   color: white;
   font-weight: bold;
-}
-
-.nav {
-  text-align: right;
 }
