+++ hike-quest-app/src/App.css
@@ -2,8 +2,4 @@
   max-width: 1280px;
   margin: 0 auto;
   padding: 2rem;
-  text-align: right;
 }
-
-
-
