+++ hike-quest-app/src/App.jsx
@@ -6,6 +6,7 @@ import Footer from './views/Footer/Footer'
 import MyPosts from './views/MyPosts/MyPosts'
 import Topics from './views/Topics/Topics'
 import Login from './views/Login/Login'
+import Post from './components/Post/Post'
 import {AppContext} from '././state/app.context'
 import { useState } from 'react'
 
@@ -21,6 +22,7 @@ function App() {
           <Route path='/account-user' element={<Account />} />
           <Route path='/my-posts' element={<MyPosts />} />
           <Route path='/login' element={<Login />} />
+          <Route path='/post' element={<Post />} />
         </Routes>
         <Footer />
       </AppContext.Provider>
