--- hike-quest-app/src/components/Post.css
@@ -1,37 +1,0 @@
-.userContainer {
-    display: flex;
-    justify-content: left;
-}
-
-.postContainer {
-    border-style: solid;
-    border-radius: 10px;
-    border-color: black;
-}
-
-.actualPost {
-    display: flex;
-    justify-content: center;
-    margin: 20px;
-    padding: 20px;
-    border-style: solid;
-    border-radius: 10px;
-    border-color: green;
-}
-
-.postButton {
-    margin: 5px;
-}
-
-.buttonContainer {
-    margin-left: 10px;
-    margin-bottom: 10px;
-    display: flex;
-    justify-content: left;
-}
-
-.userName {
-    display: flex;
-    justify-content: left;
-    margin-left: 20px;
-}
