--- hike-quest-app/src/components/Post.jsx
@@ -1,30 +1,0 @@
-import React from "react"
-import "./Post.css"
-
-
-const Post = () => {
-
-    return (
-        <>
-        
-        <div className="postContainer">
-            <div className="userContainer">
-                <img 
-                src="/Users/ivo/projects/hike-quest/MV5BNWVmZTI5M2MtNGJjYS00OWVmLTk2MDUtNmM0MTUyYmUwNWZmXkEyXkFqcGdeQXVyMjkxNzQ1NDI@._V1_.jpg" 
-                alt="profile-pic"/>
-                <a className="userName">Username</a>
-            </div>
-            <a className="actualPost">The hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.he hike around Everest was an amazing Journey. Made me feel like a god over the world when I got to the peek.</a>
-            <div className="buttonContainer">
-                <button className="postButton">Like</button>
-                <button className="postButton">Comment</button>
-            </div>
-        </div>
-
-        </>
-    )
-
-}
-
-export default Post;
-
