+++ hike-quest-app/src/components/Post/Post.css
@@ -1,0 +1,84 @@
+.userContainer {
+    display: flex;
+    justify-content: left;
+}
+
+
+
+.userName {
+    display: contents;
+    justify-content: left;
+    margin: 0;
+}
+
+.userType {
+    margin: 0;
+    font-size: 12px;
+}
+
+.profilePic {
+    margin: 10px;
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    border: 1px solid #000;
+    object-fit: cover;
+}
+
+.postContainer {
+    border-style: solid;
+    border-radius: 10px;
+    border-color: black;
+}
+
+.actualPost {
+    display: flex;
+    justify-content: center;
+    margin: 20px;
+    padding: 20px;
+    border-style: solid;
+    border-radius: 10px;
+    border-color: rgb(0, 0, 0);
+}
+
+.postButton {
+    margin: 5px;
+}
+
+.buttonContainer {
+    margin-left: 10px;
+    margin-bottom: 10px;
+    display: flex;
+    justify-content: left;
+}
+
+.commentsHeader {
+    text-align: center;
+}
+
+.commentSection {
+    display: block;
+    margin: 20px;
+    padding: 20px;
+    border-style: hidden;
+    border-radius: 10px;
+    border-color: rgb(0, 0, 0);
+}
+
+.actualComment {
+    display: block;
+    text-align: left;
+    margin: 20px 0;
+    padding: 20px;
+    border-style: dashed;
+    border-radius: 10px;
+    border-color: rgb(0, 0, 0);
+    word-wrap: break-word;
+    overflow-wrap: break-word;
+
+}
+
+hr {
+    width: 50%;
+    margin: 0 auto;
+}
