+++ hike-quest-app/src/views/MyPosts/MyPosts.jsx
@@ -1,7 +1,11 @@
+import Post from "../../components/Post/Post";
+
 export default function MyPosts() {
     return (
         <div style={{
             fontSize: '16px', fontWeight: 'bold',
-            color: '#273f17',   display: 'flex',}} >My posts</div>
+            color: '#273f17',   display: 'flex',}} >My posts
+        </div>
+
     )
 }
