+++ hike-quest-app/src/views/Topics/Topics.jsx
@@ -2,6 +2,7 @@ export default function Topics() {
     return (
         <div style={{
             fontSize: '16px', fontWeight: 'bold',
-            color: '#273f17',   display: 'flex',}}>Topics</div>
+            color: '#273f17',   display: 'flex',}}>Topics
+        </div>
     )
 }
